refactor(step1): applying feedback.
