+++ src/main/java/stringcalculator/StringCalculator.java
@@ -5,6 +5,13 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class StringCalculator {
+
+    private final static Pattern CALCULATOR_DELIMITER_PATTERN = Pattern.compile("//(.)\n(.*)");
+    public static final int CUSTOM_DELIMITER = 1;
+    public static final int USER_INPUT = 2;
+    public static final String DEFAULT_DELIMITER = "[,:]";
+    public static final String CUSTOM_DELIMITER_FORMAT = "[%s]";
+
     public int add(String text) {
         String[] inputValues = splitAndValidate(text);
 
@@ -16,21 +23,22 @@ public class StringCalculator {
             return new String[0];
         }
 
-        String[] values;
-        Matcher m = Pattern.compile("//(.)\n(.*)").matcher(text);
+        String[] userInputValuesForCalculation;
+        Matcher m = CALCULATOR_DELIMITER_PATTERN.matcher(text);
 
         if (m.find()) {
-            String customDelimiter = m.group(1);
-            values = m.group(2).split(customDelimiter);
+            String customDelimiter = String.format(CUSTOM_DELIMITER_FORMAT
+                                                ,m.group(CUSTOM_DELIMITER));
+            userInputValuesForCalculation = m.group(USER_INPUT).split(customDelimiter);
         } else {
-            values = text.split("[,:]");
+            userInputValuesForCalculation = text.split(DEFAULT_DELIMITER);
         }
-        return values;
+        return userInputValuesForCalculation;
     }
 
     private static int sum(String[] values) {
         return Arrays.stream(values)
-                .mapToInt(val -> Number.of(val).number())
+                .mapToInt(val -> OnlyPositiveNumber.of(val).number())
                 .sum();
     }
 }
